+++ src/schedule/schedule.module.ts
@@ -2,10 +2,12 @@ import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Schedule } from './schedule.entity';
 import { RecurrenceGroup } from './recurrence-group.entity';
-import { ScheduleService } from './schedule.service';
+import { ScheduleService } from './service/schedule.service';
+import { ScheduleWatchService } from './service/schedule-watch.service';
+import { ScheduleRecurringService } from './service/schedule-recurring.service';
 import { ScheduleController } from './schedule.controller';
-import { ScheduleCronService } from './schedule-cron.service';
-import { ScheduleNotificationService } from './schedule-notification.service';
+import { ScheduleCronService } from './service/schedule-cron.service';
+import { ScheduleNotificationService } from './service/schedule-notification.service';
 import { ScheduleWatchController } from './schedule-watch.controller';
 import { UserModule } from '../user/user.module';
 import { TagModule } from '../tag/tag.module';
@@ -26,9 +28,11 @@ import { GoogleModule } from '../google/google.module';
   controllers: [ScheduleController, ScheduleWatchController],
   providers: [
     ScheduleService,
+    ScheduleWatchService,
+    ScheduleRecurringService,
     ScheduleCronService,
     ScheduleNotificationService,
   ],
-  exports: [ScheduleService],
+  exports: [ScheduleService, ScheduleWatchService, ScheduleRecurringService],
 })
 export class ScheduleModule {}
