+++ src/schedule/service/schedule-recurring.service.ts
@@ -1,42 +1,37 @@
 import { Inject, Injectable, Logger } from '@nestjs/common';
-import { BusinessError, ErrorCode } from '../common/errors';
+import { BusinessError, ErrorCode } from '../../common/errors';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
 import type { Cache } from 'cache-manager';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, In } from 'typeorm';
-import { Schedule, ScheduleStatus } from './schedule.entity';
-import { RecurrenceGroup } from './recurrence-group.entity';
-import { GoogleCalendarService } from '../google/google-calendar.service';
-import { UserService } from '../user/user.service';
-import { Tag } from '../tag/tag.entity';
-import { ChannelService } from '../channel/channel.service';
+import { Schedule } from '../schedule.entity';
+import { RecurrenceGroup } from '../recurrence-group.entity';
+import { GoogleCalendarService } from '../../google/google-calendar.service';
+import { UserService } from '../../user/user.service';
+import { ChannelService } from '../../channel/channel.service';
 import { WebClient } from '@slack/web-api';
 import { randomUUID } from 'crypto';
-import {
-  buildRecurringCreationBlocks,
-  buildRecurringDeleteBlocks,
-  buildRecurringUpdateBlocks,
-} from './view/schedule-recurring.view';
 import { RRule, Weekday } from 'rrule';
 import { ScheduleNotificationService } from './schedule-notification.service';
-import { CreateScheduleDto, UpdateScheduleDto } from './dto/schedule.dto';
 import {
   RecurrenceType,
   CreateRecurringEventsDto,
   UpdateRecurringEventsDto,
-} from './dto/recurring.dto';
-
+} from '../dto/recurring.dto';
+import {
+  buildRecurringCreationBlocks,
+  buildRecurringDeleteBlocks,
+  buildRecurringUpdateBlocks,
+} from '../view/schedule-recurring.view';
 
 @Injectable()
-export class ScheduleService {
-  private readonly logger = new Logger(ScheduleService.name);
+export class ScheduleRecurringService {
+  private readonly logger = new Logger(ScheduleRecurringService.name);
   private readonly slack = new WebClient(process.env.SLACK_BOT_TOKEN);
 
   constructor(
     @InjectRepository(Schedule)
     private scheduleRepository: Repository<Schedule>,
-    @InjectRepository(Tag)
-    private tagRepository: Repository<Tag>,
     @InjectRepository(RecurrenceGroup)
     private recurrenceGroupRepository: Repository<RecurrenceGroup>,
     @Inject(CACHE_MANAGER) private cache: Cache,
@@ -45,478 +40,6 @@ export class ScheduleService {
     private readonly userService: UserService,
     private readonly scheduleNotificationService: ScheduleNotificationService,
   ) {}
-
-  // 스케줄 생성 (Google Calendar도 함께 생성)
-  async createSchedule(dto: CreateScheduleDto): Promise<Schedule> {
-    // 1. Google Calendar 생성
-    const { calendarId } = await this.googleCalendarService.createCalendar(
-      dto.name,
-      dto.description,
-    );
-
-    // 2. 태그 조회
-    let tags: Tag[] = [];
-    if (dto.tagIds && dto.tagIds.length > 0) {
-      tags = await this.tagRepository.find({
-        where: { id: In(dto.tagIds) },
-      });
-    }
-
-    // 3. Schedule 엔티티 저장
-    const schedule = this.scheduleRepository.create({
-      name: dto.name,
-      calendarId,
-      description: dto.description,
-      status: ScheduleStatus.ACTIVE,
-      tags,
-      createdById: dto.createdById,
-    });
-
-    const saved = await this.scheduleRepository.save(schedule);
-
-    // 4. 캘린더 전체 공개 설정
-    await this.googleCalendarService.makeCalendarPublic(calendarId);
-
-    // 5. 생성자에게 writer 권한 부여 및 자동 구독
-    if (dto.creatorEmail && dto.creatorRefreshToken) {
-      await this.googleCalendarService.shareCalendar({
-        calendarId,
-        email: dto.creatorEmail,
-        role: 'writer',
-      });
-      await this.googleCalendarService.addCalendarToUserList(
-        calendarId,
-        dto.creatorRefreshToken,
-      );
-    }
-
-    // 6. Google Calendar Watch 등록
-    await this.registerWatch(saved.id);
-
-    return saved;
-  }
-
-  async findById(id: number): Promise<Schedule | null> {
-    return this.scheduleRepository.findOne({
-      where: { id },
-      relations: ['tags', 'createdBy'],
-    });
-  }
-
-  async findByCalendarId(calendarId: string): Promise<Schedule | null> {
-    return this.scheduleRepository.findOne({
-      where: { calendarId },
-      relations: ['tags', 'createdBy'],
-    });
-  }
-
-  async findByName(name: string): Promise<Schedule | null> {
-    return this.scheduleRepository.findOne({
-      where: { name },
-      relations: ['tags'],
-    });
-  }
-
-  // 활성 스케줄 목록 조회
-  async findActiveSchedules(): Promise<Schedule[]> {
-    return this.scheduleRepository.find({
-      where: { status: ScheduleStatus.ACTIVE },
-      relations: ['tags', 'createdBy'],
-      order: { name: 'ASC' },
-    });
-  }
-
-  // 특정 태그를 가진 활성 스케줄 목록 조회 (태그 시간표용)
-  async findActiveSchedulesByTagId(tagId: number): Promise<Schedule[]> {
-    const ids = await this.scheduleRepository
-      .createQueryBuilder('schedule')
-      .innerJoin('schedule.tags', 'tag')
-      .where('tag.id = :tagId', { tagId })
-      .andWhere('schedule.status = :status', { status: ScheduleStatus.ACTIVE })
-      .andWhere('schedule.deletedAt IS NULL')
-      .select('schedule.id')
-      .getRawMany()
-      .then((rows: { schedule_id: number }[]) =>
-        rows.map((r) => r.schedule_id),
-      );
-
-    if (ids.length === 0) return [];
-
-    return this.scheduleRepository.find({
-      where: { id: In(ids) },
-      order: { name: 'ASC' },
-    });
-  }
-
-  // 반 대표 슬랙 ID로 자기 반 태그가 달린 활성 스케줄 조회
-  async findSchedulesByClassRepSlackId(
-    slackUserId: string,
-  ): Promise<Schedule[]> {
-    const user = await this.userService.findBySlackIdWithClass(slackUserId);
-    if (!user?.studentClassId) return [];
-
-    const tag = await this.tagRepository.findOne({
-      where: { studentClassId: user.studentClassId },
-    });
-    if (!tag) return [];
-
-    return this.findActiveSchedulesByTagId(tag.id);
-  }
-
-  // 반 대표가 해당 과목에 대한 권한이 있는지 확인 (자기 반 태그 포함 여부)
-  async isClassRepAuthorizedForSchedule(
-    slackUserId: string,
-    scheduleId: number,
-  ): Promise<boolean> {
-    const user = await this.userService.findBySlackIdWithClass(slackUserId);
-    if (!user?.studentClassId) return false;
-
-    const tag = await this.tagRepository.findOne({
-      where: { studentClassId: user.studentClassId },
-    });
-    if (!tag) return false;
-
-    const count = await this.scheduleRepository
-      .createQueryBuilder('schedule')
-      .innerJoin('schedule.tags', 'tag')
-      .where('schedule.id = :scheduleId', { scheduleId })
-      .andWhere('tag.id = :tagId', { tagId: tag.id })
-      .getCount();
-
-    return count > 0;
-  }
-
-  // 태그가 없는 활성 스케줄 목록 조회
-  async findActiveSchedulesWithoutTags(): Promise<Schedule[]> {
-    return this.scheduleRepository
-      .createQueryBuilder('schedule')
-      .leftJoin('schedule.tags', 'tag')
-      .where('schedule.status = :status', { status: ScheduleStatus.ACTIVE })
-      .andWhere('schedule.deletedAt IS NULL')
-      .groupBy('schedule.id')
-      .having('COUNT(tag.id) = 0')
-      .orderBy('schedule.name', 'ASC')
-      .getMany();
-  }
-
-  // 모든 스케줄 목록 조회 (관리용)
-  async findAllSchedules(): Promise<Schedule[]> {
-    return this.scheduleRepository.find({
-      relations: ['tags', 'createdBy'],
-      order: { status: 'ASC', name: 'ASC' },
-    });
-  }
-
-  // 페이지네이션 + 필터 조회 (관리용)
-  async findSchedulesPaginated(opts: {
-    page: number;
-    pageSize: number;
-    status?: 'active' | 'inactive';
-    tagIds?: number[];
-  }): Promise<{ schedules: Schedule[]; total: number }> {
-    const { page, pageSize, status, tagIds } = opts;
-
-    if (tagIds && tagIds.length > 0) {
-      // 태그 필터: 선택한 태그를 모두 포함하는 스케줄 (AND 조건)
-      const idQb = this.scheduleRepository
-        .createQueryBuilder('schedule')
-        .innerJoin('schedule.tags', 'tag')
-        .where('tag.id IN (:...tagIds)', { tagIds })
-        .andWhere('schedule.deletedAt IS NULL');
-
-      if (status) {
-        idQb.andWhere('schedule.status = :status', {
-          status:
-            status === 'active'
-              ? ScheduleStatus.ACTIVE
-              : ScheduleStatus.INACTIVE,
-        });
-      }
-
-      const ids = (
-        await idQb
-          .groupBy('schedule.id')
-          .having('COUNT(DISTINCT tag.id) = :tagCount', {
-            tagCount: tagIds.length,
-          })
-          .select('schedule.id')
-          .getRawMany()
-      ).map((r: { schedule_id: number }) => r.schedule_id);
-
-      if (ids.length === 0) return { schedules: [], total: 0 };
-
-      const [schedules, total] = await this.scheduleRepository.findAndCount({
-        where: { id: In(ids) },
-        relations: ['tags', 'createdBy'],
-        order: { status: 'ASC', name: 'ASC' },
-        skip: page * pageSize,
-        take: pageSize,
-      });
-
-      return { schedules, total };
-    }
-
-    // 태그 필터 없음 → 기본 find
-    const where: Record<string, unknown> = {};
-    if (status) {
-      where['status'] =
-        status === 'active' ? ScheduleStatus.ACTIVE : ScheduleStatus.INACTIVE;
-    }
-
-    const [schedules, total] = await this.scheduleRepository.findAndCount({
-      where,
-      relations: ['tags', 'createdBy'],
-      order: { status: 'ASC', name: 'ASC' },
-      skip: page * pageSize,
-      take: pageSize,
-    });
-
-    return { schedules, total };
-  }
-
-  // 스케줄 업데이트 (Google Calendar도 함께 업데이트)
-  async updateSchedule(
-    id: number,
-    dto: UpdateScheduleDto,
-  ): Promise<Schedule | null> {
-    const schedule = await this.findById(id);
-    if (!schedule) return null;
-
-    // Google Calendar 업데이트
-    if (dto.name || dto.description !== undefined) {
-      await this.googleCalendarService.updateCalendar(
-        schedule.calendarId,
-        dto.name ?? schedule.name,
-        dto.description ?? schedule.description,
-      );
-    }
-
-    // 태그 업데이트
-    if (dto.tagIds) {
-      const tags = await this.tagRepository.find({
-        where: { id: In(dto.tagIds) },
-      });
-      schedule.tags = tags;
-    }
-
-    if (dto.name) schedule.name = dto.name;
-    if (dto.description !== undefined) schedule.description = dto.description;
-
-    return this.scheduleRepository.save(schedule);
-  }
-
-  // 스케줄 비활성화
-  async deactivateSchedule(id: number): Promise<Schedule | null> {
-    await this.stopWatch(id);
-    await this.scheduleRepository.update(
-      { id },
-      { status: ScheduleStatus.INACTIVE },
-    );
-    return this.findById(id);
-  }
-
-  // 스케줄 활성화
-  async activateSchedule(id: number): Promise<Schedule | null> {
-    await this.scheduleRepository.update(
-      { id },
-      { status: ScheduleStatus.ACTIVE },
-    );
-    await this.registerWatch(id);
-    return this.findById(id);
-  }
-
-  // 스케줄 삭제 (soft delete, Google Calendar도 삭제)
-  async deleteSchedule(id: number): Promise<void> {
-    const schedule = await this.findById(id);
-    if (!schedule) return;
-
-    await this.stopWatch(id);
-
-    // Google Calendar 삭제
-    try {
-      await this.googleCalendarService.deleteCalendar(schedule.calendarId);
-    } catch (error) {
-      // Calendar가 이미 삭제된 경우 무시
-      this.logger.warn(
-        `Failed to delete calendar ${schedule.calendarId}: ${error}`,
-      );
-    }
-
-    await this.scheduleRepository.softDelete({ id });
-  }
-
-  // ========== Google Calendar Watch ==========
-
-  // watch 등록 — 멱등 (기존 watch가 있으면 stop 후 재등록)
-  async registerWatch(id: number): Promise<void> {
-    const schedule = await this.scheduleRepository.findOne({ where: { id } });
-    if (!schedule) return;
-
-    if (!this.googleCalendarService.isWatchConfigured()) {
-      this.logger.warn(
-        'GOOGLE_WEBHOOK_URL not set, skipping watch registration',
-      );
-      return;
-    }
-
-    // 기존 watch가 있으면 먼저 해제
-    if (schedule.watchChannelId && schedule.watchResourceId) {
-      try {
-        await this.googleCalendarService.stopCalendarWatch(
-          schedule.watchChannelId,
-          schedule.watchResourceId,
-        );
-      } catch (error) {
-        this.logger.warn(
-          `Failed to stop existing watch for schedule ${id}: ${error}`,
-        );
-      }
-    }
-
-    const channelId = randomUUID();
-
-    try {
-      const { resourceId } =
-        await this.googleCalendarService.watchCalendarEvents(
-          schedule.calendarId,
-          channelId,
-        );
-
-      const syncToken = await this.googleCalendarService.getInitialSyncToken(
-        schedule.calendarId,
-      );
-
-      await this.scheduleRepository.update(id, {
-        watchChannelId: channelId,
-        watchResourceId: resourceId,
-        syncToken,
-      });
-
-      this.logger.log(
-        `Watch registered for schedule ${id} (channelId: ${channelId})`,
-      );
-    } catch (error) {
-      this.logger.error(
-        `Failed to register watch for schedule ${id}: ${error}`,
-      );
-    }
-  }
-
-  // watch 해제
-  async stopWatch(id: number): Promise<void> {
-    const schedule = await this.scheduleRepository.findOne({ where: { id } });
-    if (!schedule?.watchChannelId || !schedule?.watchResourceId) return;
-
-    try {
-      await this.googleCalendarService.stopCalendarWatch(
-        schedule.watchChannelId,
-        schedule.watchResourceId,
-      );
-    } catch (error) {
-      this.logger.warn(`Failed to stop watch for schedule ${id}: ${error}`);
-    }
-
-    await this.scheduleRepository.update(id, {
-      watchChannelId: null,
-      watchResourceId: null,
-    });
-
-    this.logger.log(`Watch stopped for schedule ${id}`);
-  }
-
-  // 모든 active 스케줄 watch 일괄 갱신 (weekly cron 전용)
-  async renewAllActiveWatches(): Promise<void> {
-    const activeSchedules = await this.scheduleRepository.find({
-      where: { status: ScheduleStatus.ACTIVE },
-      select: ['id'],
-    });
-
-    this.logger.log(
-      `Renewing watches for ${activeSchedules.length} active schedules`,
-    );
-
-    for (const { id } of activeSchedules) {
-      await this.registerWatch(id);
-    }
-  }
-
-  // watchChannelId로 스케줄 조회 (웹훅 수신 시 사용)
-  async findByWatchChannelId(channelId: string): Promise<Schedule | null> {
-    return this.scheduleRepository.findOne({
-      where: { watchChannelId: channelId },
-    });
-  }
-
-  // syncToken 업데이트 (웹훅 수신 후 새 토큰 저장)
-  async updateSyncToken(id: number, syncToken: string): Promise<void> {
-    await this.scheduleRepository.update(id, { syncToken });
-  }
-
-  // 캘린더 권한 목록 조회 (Google Calendar API 사용)
-  async getCalendarPermissions(id: number) {
-    const schedule = await this.findById(id);
-    if (!schedule) return null;
-
-    return this.googleCalendarService.getCalendarAcl(schedule.calendarId);
-  }
-
-  // 캘린더 권한 부여
-  async shareCalendar(
-    id: number,
-    email: string,
-    role: 'reader' | 'writer' | 'owner',
-  ): Promise<void> {
-    const schedule = await this.findById(id);
-    if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
-
-    await this.googleCalendarService.shareCalendar({
-      calendarId: schedule.calendarId,
-      email,
-      role,
-    });
-  }
-
-  // 캘린더 권한 제거
-  async unshareCalendar(id: number, email: string): Promise<void> {
-    const schedule = await this.findById(id);
-    if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
-
-    await this.googleCalendarService.unshareCalendar(
-      schedule.calendarId,
-      email,
-    );
-  }
-
-  // 구독 (사용자 캘린더 목록에 추가)
-  async subscribe(id: number, userRefreshToken: string): Promise<void> {
-    const schedule = await this.findById(id);
-    if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
-
-    await this.googleCalendarService.addCalendarToUserList(
-      schedule.calendarId,
-      userRefreshToken,
-    );
-  }
-
-  // 구독 해제 (사용자 캘린더 목록에서 제거)
-  async unsubscribe(id: number, userRefreshToken: string): Promise<void> {
-    const schedule = await this.findById(id);
-    if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
-
-    await this.googleCalendarService.removeCalendarFromUserList(
-      schedule.calendarId,
-      userRefreshToken,
-    );
-  }
-
-  // 사용자의 구독 중인 calendarId Set 반환 (모달 빌드 시 일괄 비교용)
-  async getSubscribedCalendarIds(
-    userRefreshToken: string,
-  ): Promise<Set<string>> {
-    return this.googleCalendarService.getUserCalendarIds(userRefreshToken);
-  }
-
-  // ========== 반복 일정 ==========
 
   private async resolveWriterDisplay(calendarId: string): Promise<string> {
     try {
@@ -533,11 +56,15 @@ export class ScheduleService {
     }
   }
 
+  // 반복 일정 생성
   async createRecurringEvents(
     dto: CreateRecurringEventsDto,
     executorSlackId?: string,
   ): Promise<void> {
-    const schedule = await this.findById(dto.scheduleId);
+    const schedule = await this.scheduleRepository.findOne({
+      where: { id: dto.scheduleId },
+      relations: ['tags', 'createdBy'],
+    });
     if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
 
     // 1. 날짜 배열 계산
@@ -550,18 +77,20 @@ export class ScheduleService {
     await this.cache.set(`suppress:group:${groupId}`, true, 3 * 60 * 1000);
 
     // 3. 이벤트 10개씩 청크 생성 (Rate Limit 방지)
-    const results = await runInChunks(dates, ({ startDateTime, endDateTime }) =>
-      this.googleCalendarService.createEventAsServiceAccount(
-        schedule.calendarId,
-        {
-          summary: dto.title,
-          startDateTime,
-          endDateTime,
-          description: dto.description,
-          location: dto.location,
-          groupId,
-        },
-      ),
+    const results = await runInChunks(
+      dates,
+      ({ startDateTime, endDateTime }) =>
+        this.googleCalendarService.createEventAsServiceAccount(
+          schedule.calendarId,
+          {
+            summary: dto.title,
+            startDateTime,
+            endDateTime,
+            description: dto.description,
+            location: dto.location,
+            groupId,
+          },
+        ),
     );
 
     const successCount = results.filter((r) => r.status === 'fulfilled').length;
@@ -596,7 +125,7 @@ export class ScheduleService {
       recurrenceType: dto.recurrenceType,
     });
 
-    // 5. Slack 채널에 요약 알림 직접 발송 (웹훅 알림 차단용 auto-mute 선행)
+    // 5. Slack 채널에 요약 알림 발송 (웹훅 알림 차단용 auto-mute 선행)
     await this.scheduleNotificationService.autoMute(dto.scheduleId);
 
     const slackChannelIds = await this.channelService.getSlackChannelIds(
@@ -636,11 +165,12 @@ export class ScheduleService {
     );
   }
 
+  // id로 반복 그룹 조회
   async findRecurrenceGroupById(id: number): Promise<RecurrenceGroup | null> {
     return this.recurrenceGroupRepository.findOne({ where: { id } });
   }
 
-  // groupId로 그룹 조회
+  // 스케줄 ID로 반복 그룹 목록 조회
   async findRecurrenceGroupsBySchedule(
     scheduleId: number,
   ): Promise<RecurrenceGroup[]> {
@@ -683,6 +213,7 @@ export class ScheduleService {
     return schedules.map((s) => ({ id: s.id, name: s.name }));
   }
 
+  // 반복 그룹 삭제
   async deleteRecurringGroup(
     groupDbId: number,
     scope: 'all' | 'future',
@@ -694,7 +225,10 @@ export class ScheduleService {
     });
     if (!group) throw new BusinessError(ErrorCode.RECURRENCE_GROUP_NOT_FOUND);
 
-    const schedule = await this.findById(group.scheduleId);
+    const schedule = await this.scheduleRepository.findOne({
+      where: { id: group.scheduleId },
+      relations: ['tags', 'createdBy'],
+    });
     if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
 
     await this.cache.set(
@@ -779,6 +313,7 @@ export class ScheduleService {
     return { deleted: deletedCount, total: events.length };
   }
 
+  // 반복 그룹 수정
   async updateRecurringGroup(
     groupDbId: number,
     dto: UpdateRecurringEventsDto,
@@ -790,7 +325,10 @@ export class ScheduleService {
     });
     if (!group) throw new BusinessError(ErrorCode.RECURRENCE_GROUP_NOT_FOUND);
 
-    const schedule = await this.findById(group.scheduleId);
+    const schedule = await this.scheduleRepository.findOne({
+      where: { id: group.scheduleId },
+      relations: ['tags', 'createdBy'],
+    });
     if (!schedule) throw new BusinessError(ErrorCode.SCHEDULE_NOT_FOUND);
 
     await this.cache.set(
@@ -822,7 +360,7 @@ export class ScheduleService {
       dto.endDate !== undefined;
 
     if (needsRecreate) {
-      // ── 요일/기간 변경: 기존 원본 전체 삭제 후 재생성 ───────────────
+      // 요일/기간 변경: 기존 원본 전체 삭제 후 재생성
       await runInChunks(events, (e) =>
         this.googleCalendarService.deleteEventAsServiceAccount(
           schedule.calendarId,
@@ -870,7 +408,7 @@ export class ScheduleService {
         ).length;
       }
     } else {
-      // ── 일반 수정 (제목·시각·설명·장소) ──────────────────────────────
+      // 일반 수정 (제목·시각·설명·장소)
       const results = await runInChunks(events, (e) => {
         let startDateTime: string | undefined;
         let endDateTime: string | undefined;
@@ -899,7 +437,7 @@ export class ScheduleService {
       updatedCount = results.filter((r) => r.status === 'fulfilled').length;
     }
 
-    // DB 업데이트 데이터 빌드
+    // DB 업데이트
     const updateData: Partial<RecurrenceGroup> = {};
     if (dto.title) updateData.title = dto.title;
     if (dto.location !== undefined) updateData.location = dto.location;
@@ -961,13 +499,11 @@ export class ScheduleService {
 }
 
 // ========== 날짜 확장 헬퍼 ==========
-// TODO 타입 의존성 제거 후 DateUtil로 이동 예정
 
 function expandRecurringDates(dto: CreateRecurringEventsDto): {
   startDateTime: string;
   endDateTime: string;
 }[] {
-  // UTC midnight으로 생성 → rrule이 UTC 기준 날짜를 그대로 반환
   const dtstart = new Date(`${dto.startDate}T00:00:00Z`);
   const until = new Date(`${dto.endDate}T00:00:00Z`);
 
@@ -986,7 +522,6 @@ function expandRecurringDates(dto: CreateRecurringEventsDto): {
   const pad = (n: number) => String(n).padStart(2, '0');
 
   return rule.all().map((date) => {
-    // getUTC* 사용으로 서버 timezone 무관하게 정확한 날짜 추출
     const y = date.getUTCFullYear();
     const mo = pad(date.getUTCMonth() + 1);
     const d = pad(date.getUTCDate());
@@ -1057,7 +592,6 @@ async function runInChunks<T, R>(
 
 // ========== 원본 이벤트 판별 헬퍼 ==========
 
-/** Google Calendar 이벤트 dateTime 문자열에서 KST 기준 HH:MM 추출 */
 function extractTimeKST(dateTimeStr: string): string {
   const d = new Date(dateTimeStr);
   const kstHours = (d.getUTCHours() + 9) % 24;
@@ -1065,17 +599,12 @@ function extractTimeKST(dateTimeStr: string): string {
   return `${String(kstHours).padStart(2, '0')}:${String(kstMinutes).padStart(2, '0')}`;
 }
 
-/** Google Calendar 이벤트 dateTime 문자열에서 KST 기준 요일(0=일 … 6=토) 추출 */
 function getDayOfWeekKST(dateTimeStr: string): number {
   const d = new Date(dateTimeStr);
   const kstMs = d.getTime() + 9 * 60 * 60 * 1000;
   return new Date(kstMs).getUTCDay();
 }
 
-/**
- * 이벤트가 반복 그룹의 원본 생성 조건(제목·시작시각·종료시각·요일)을 만족하는지 확인.
- * 하나라도 다르면 개별 수정된 것으로 판단해 false를 반환한다.
- */
 function isOriginalEvent(
   event: {
     summary?: string | null;
